upgraded, compatible with OH 2.5.0~M2-1
Secure Send enabled

+++ org.openhab.binding.souliss/src/main/java/org/openhab/binding/souliss/handler/SoulissGatewayJobHealty.java
@@ -25,8 +25,8 @@ public class SoulissGatewayJobHealty extends Thread {
 
     private Logger logger = LoggerFactory.getLogger(SoulissGatewayJobHealty.class);
     private String _iPAddressOnLAN;
-    private short _userIndex;
-    private short _nodeIndex;
+    private byte _userIndex;
+    private byte _nodeIndex;
     private int _healthRefreshInterval;
 
     private SoulissGatewayHandler gw;
