+++ org.openhab.binding.souliss/src/main/java/org/openhab/binding/souliss/handler/SoulissGatewayJobSubscription.java
@@ -23,8 +23,8 @@ public class SoulissGatewayJobSubscription extends Thread {
 
     private Logger logger = LoggerFactory.getLogger(SoulissGatewayJobSubscription.class);
     private String _iPAddressOnLAN;
-    private short _userIndex;
-    private short _nodeIndex;
+    private byte _userIndex;
+    private byte _nodeIndex;
     private int _subscriptionRefreshInterval;
 
     private SoulissGatewayHandler gw;
